Assignments/pyproject.toml
[tool.poetry]
name = "modern-software-dev-assignments"
version = "0.1.0"
description = "Course assignments with FastAPI + SQLite starter and tooling"
authors = ["Course <course@example.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.111.0"
uvicorn = { version = ">=0.23.0", extras = ["standard"] }
sqlalchemy = ">=2.0.0"
pydantic = ">=2.0.0"
python-dotenv = ">=1.0.0"
openai = ">=1.0.0"
ollama = "^0.5.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
httpx = ">=0.24.0"
black = ">=24.1.0"
ruff = ">=0.4.0"
pre-commit = ">=3.6.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]
